minor update; update run.sh script

+++ cm/karras_diffusion.py
@@ -377,7 +377,7 @@ def karras_sample(
         sigmas = get_sigmas_karras(steps + 1, sigma_min, sigma_max, rho, device=device)
     else:
         sigmas = get_sigmas_karras(steps, sigma_min, sigma_max, rho, device=device)
-    x_T = generator.randn(*shape, device=device).half() * sigma_max
+    x_T = generator.randn(*shape, device=device) * sigma_max
 
     sample_fn = {
         "heun": sample_heun,
